Fix I2C_ADDRESS sign warning (#19685)

+++ Marlin/src/HAL/HAL.h
@@ -34,7 +34,7 @@
 #define HAL_ADC_RANGE _BV(HAL_ADC_RESOLUTION)
 
 #ifndef I2C_ADDRESS
-  #define I2C_ADDRESS(A) (A)
+  #define I2C_ADDRESS(A) uint8_t(A)
 #endif
 
 // Needed for AVR sprintf_P PROGMEM extension
